fixes to cutout to make them more consistent, as well as psf_sky in correct place (with cutout where PSF should be created)
